+++ src/main/java/com/acltabontabon/openwealth/services/customermgmt/contact/ContactDeleter.java
@@ -27,7 +27,11 @@ public class ContactDeleter extends DeleteCommand<Result<Void>> {
         try {
             restClient.delete()
                 .uri(builder -> builder.path(apiProperties.getPersonContact()).build(customerId, personId, contactId))
-                .header(HEADER_CORRELATION_ID, correlationId)
+                .headers(headers -> {
+                    if (correlationId != null) {
+                        headers.set(HEADER_CORRELATION_ID, correlationId);
+                    }
+                })
                 .retrieve()
                 .toBodilessEntity();