+++ src/main/java/com/acltabontabon/openwealth/services/customermgmt/contact/SingleContactReader.java
@@ -28,7 +28,11 @@ public class SingleContactReader extends ReadCommand<Result<Contact>> {
         try {
             Contact contact = restClient.get()
                 .uri(builder -> builder.path(apiProperties.getPersonContact()).build(customerId, personId, contactId))
-                .header(HEADER_CORRELATION_ID, correlationId)
+                .headers(headers -> {
+                    if (correlationId != null) {
+                        headers.set(HEADER_CORRELATION_ID, correlationId);
+                    }
+                })
                 .retrieve()
                 .body(Contact.class);