+++ src/main/java/com/acltabontabon/openwealth/services/customermgmt/customer/SingleCustomerReader.java
@@ -61,7 +61,11 @@ public class SingleCustomerReader extends ReadCommand<Result<Customer>> {
                         return builder.path(apiProperties.getCustomer()).build(customerId);
                     }
                 })
-                .header(HEADER_CORRELATION_ID, correlationId)
+                .headers(headers -> {
+                    if (correlationId != null) {
+                        headers.set(HEADER_CORRELATION_ID, correlationId);
+                    }
+                })
                 .retrieve()
                 .body(Customer.class);